Modify and add tests to Lottery.test
Test for contract deployment largely unchanged from Inbox project.
New test checks that a single account is allowed to enter with proper
amount of ether provided. New test checks that multiple accounts are
allowed to enter with proper amount of ether provided. Both tests
verify that numbers of accounts stored is equal to number of players.